+++ src/minecraft/com/vbitz/MinecraftScript/MinecraftScriptClassShutter.java
@@ -16,6 +16,7 @@ public class MinecraftScriptClassShutter implements ClassShutter {
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptScriptedBlockAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptScriptedItemAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptNBTagAPI");
+		_allowedClasses.add("com.vbitz.MinecraftScript.MinecraftScriptNodeAPI");
 		_allowedClasses.add("com.vbitz.MinecraftScript.Vector3f");
 		_allowedClasses.add("com.vbitz.MinecraftScript.MCSCollection");
 		_allowedClasses.add("java.lang.String");
@@ -29,4 +30,5 @@ public class MinecraftScriptClassShutter implements ClassShutter {
 		} else {
 			return _allowedClasses.contains(className);
 		}
-	}}
+	}
+}