+++ src/minecraft/com/vbitz/MinecraftScript/survival/SurvivalControl.java
@@ -1,0 +1,11 @@
+package com.vbitz.MinecraftScript.survival;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SurvivalControl {
+	int cost();
+}